EarthSpells, SpellDisplay, Jumping
Adjusted EarthSpells
Fixed bug - Spell not showing in UI
Fixed bug - wall jumping

+++ Assets/Scripts/UIScripts/SpellCooldown.cs
@@ -56,7 +56,7 @@ public class SpellCooldown : MonoBehaviour
         
     }
     
-    public void UseSpell(int cooldown)
+    public void UseSpell(float cooldown)
     {
         cooldownTime = cooldown;
         isCooldown = true;
